fix(adguardhome-sync): deprecated variable update, boolean values (#22694)
**Description**
See deprecation notice here:
[https://github.com/bakito/adguardhome-sync/wiki/Deprecations#deprecation-of-environment-variables-as-of-v060](https://github.com/bakito/adguardhome-sync/wiki/Deprecations#deprecation-of-environment-variables-as-of-v060)

Also change boolean values to bool instead of string, hopefully fixes
this error when running `heavyscript update` on Scale:
```
adguardhome-sync
Failed to update. Manual intervention may be required.
Validation errors:
* values.workload.main.podSpec.containers.main.env.FEATURES_CLIENTSETTINGS: Not a boolean
* values.workload.main.podSpec.containers.main.env.FEATURES_DHCP_SERVERCONFIG: Not a boolean
* values.workload.main.podSpec.containers.main.env.FEATURES_DHCP_STATICLEASES: Not a boolean
* values.workload.main.podSpec.containers.main.env.FEATURES_DNS_ACCESSLISTS: Not a boolean
* values.workload.main.podSpec.containers.main.env.FEATURES_DNS_REWRITES: Not a boolean
* values.workload.main.pod...
For full error details, visit TrueNAS SCALE GUI: Check the Jobs icon and select your failed job.
```

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [x] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
It has not. If there is a good way for me to test, let me know and I
will.

**📃 Notes:**
Not sure if changing variable names is breaking, will people have to
reinput them? If so close and I can bump minor version instead of patch.
Or if there's a way to direct the chart how to rename current values let
me know and I'll do that.

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning
- [x] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):` or `chore(chart-name):`

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: Morgan Hunter <32530812+mxc42@users.noreply.github.com>