fix(prowlarr): Change prowlarr exportarr pod into a sidecar container (#25111)
**Description**

Prowlarr helm chart and other Arrs implementing a exportarr pod can fail
when deployed on a multi node cluster when the pods are provisioned on
different nodes. This is because both mount the config PVC which is in
ReadWriteOnce accessModes by default which only allows mounting on the
same node.

Fixes #25130

The solution I chose is to convert the exportarr pod into a sidecar
container. I choose this approach for simplicity.

If this is not acceptable I will gladly change the approach and create a
new PR and create one for each other affected Arr chart.

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [x] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**

Deployed on my home lab.

**📃 Notes:**
<!-- Please enter any other relevant information here -->

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [ ] ⬆️ I increased versions for any altered app according to semantic
versioning
- [x] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):` or `chore(chart-name):`

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: alfi0812 <43101280+alfi0812@users.noreply.github.com>
Co-authored-by: alfi0812 <43101280+alfi0812@users.noreply.github.com>