chore(clustertool) change included patch format and add nvidia patch (#31813)
**Description**
Change clustertool included patches from RFC6902 (json patches) to
strategic merge patches, see
[talos-guides](https://www.talos.dev/v1.9/talos-guides/configuration/patching/).
This allows for the used to easily add to some of the "list" values like
sysctls and kernel modules.

For example, this list operation overwrites any previous adds to
sysctls.
```yaml
- op: add
  path: /machine/sysctls
  value:
    fs.inotify.max_queued_events: "65536"
    fs.inotify.max_user_instances: "8192"
```

this is not the case when using strategic merge patches
```yaml
machine:
  sysctls:
    fs.inotify.max_queued_events: "65536"
    fs.inotify.max_user_instances: "8192"
```

for example nvidia.yaml can add to this
```yaml
machine:
  sysctls:
    net.core.bpf_jit_harden: 1
```
⚒️ Fixes  # <!--(issue)-->

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [x] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
I ran clustertool genconfig with the default configs before and after
the change. The resulting main-k8s-control-1.yaml was identical (except,
see below).

**📃 Notes:**

The bind addresses for proxy and scheduler did never end up in the
generated config because they got overwritten by subsequent patch
operations. Now these entries will be in the generated config, not sure
if that could be an issue?

**✔️ Checklist:**

- [ ] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [ ] ⬆️ I increased versions for any altered app according to semantic
versioning
- [x] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):` or `chore(chart-name):`

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._