Add documentation for `clustertool talos kubeconfig` command (#32159)
**Description**
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->
Added the documentation to the `clustertool talos kubeconfig`, I needed
to use this command and I saw it in the help page `clustertool talos -h`
and source code, but not in the documentation online that I was using as
reference. I tried to add the required page
⚒️ Fixes  # <!--(issue)-->
No fixes made
**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->
I checked the help page of the command `clustertool talos kubeconfig`

**📃 Notes:**
<!-- Please enter any other relevant information here -->
I never programmed in Go, but i wanted to check in the source code to be
sure to not write a bad documentation. It seems to me that this command
is just a wrapper around the `talosctl kubeconfig -n <NodeIP>`, but it
seams to use a function called `gencmd.GenPlain(...)` that seems to wrap
the talosctl command verbatim. If it does it the help page should be
different and also the section related to the option in my
documentation.

![250216-152545_screenshot-area](https://github.com/user-attachments/assets/9a768f3e-4219-4953-8c90-91f2993be385)

![250216-152442_screenshot-area](https://github.com/user-attachments/assets/5d3e4557-4585-4219-9d0d-31b54fec3acb)

**✔️ Checklist:**

- [ ] ⚖️ My code follows the style guidelines of this project
- [ ] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [x] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [ ] ⬆️ I increased versions for any altered app according to semantic
versioning
- [ ] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):` or `chore(chart-name):`

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._